# core-api/legacy_scripts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financial Analysis API",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Production server (update with actual URL)"
    }
  ],
  "paths": {
...
    "/api/import": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Transactions",
        "description": "Import transactions from Excel file.\n\nArgs:\n    file: Uploaded Excel file\n    mode: Import mode (full, incremental, update)\n    skip_duplicates: Whether to skip duplicates\n    db: Database session\n    \nReturns:\n    ImportResponse: Import result",
        "operationId": "import_transactions_api_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_transactions_api_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/file-path": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import From File Path",
        "description": "Import transactions from file path (for local development).\n\nArgs:\n    file_path: Path to Excel file\n    request: Import request parameters\n    db: Database session\n    \nReturns:\n    ImportResponse: Import result",
        "operationId": "import_from_file_path_api_import_file_path_post",
        "parameters": [
          {
            "name": "file_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_transactions_api_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Excel file to import"
          },
          "mode": {
            "type": "string",
            "pattern": "^(full|incremental|update)$",
            "title": "Mode",
            "description": "Import mode",
            "default": "incremental"
          },
          "skip_duplicates": {
            "type": "boolean",
            "title": "Skip Duplicates",
            "description": "Skip duplicate transactions",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_transactions_api_import_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportErrorDetail": {
        "properties": {
          "row": {
            "type": "integer",
            "title": "Row",
            "description": "Row number"
          },
          "field": {
            "type": "string",
            "title": "Field",
            "description": "Field name"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Error message"
          },
          "value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "Invalid value"
          }
        },
        "type": "object",
        "required": [
          "row",
          "field",
          "message"
        ],
        "title": "ImportErrorDetail",
        "description": "Schema for import error detail."
      },
      "ImportRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "pattern": "^(full|incremental|update)$",
            "title": "Mode",
            "description": "Import mode",
            "default": "incremental"
          },
          "skip_duplicates": {
            "type": "boolean",
            "title": "Skip Duplicates",
            "description": "Skip duplicate transactions",
            "default": true
          }
        },
        "type": "object",
        "title": "ImportRequest",
        "description": "Schema for import request."
      },
      "ImportResponse": {
        "properties": {
          "total_rows": {
            "type": "integer",
            "title": "Total Rows",
            "description": "Total rows processed"
          },
          "successful_rows": {
            "type": "integer",
            "title": "Successful Rows",
            "description": "Successfully imported rows"
          },
          "failed_rows": {
            "type": "integer",
            "title": "Failed Rows",
            "description": "Failed rows"
          },
          "classified_rows": {
            "type": "integer",
            "title": "Classified Rows",
            "description": "Automatically classified rows"
          },
          "skipped_duplicates": {
            "type": "integer",
            "title": "Skipped Duplicates",
            "description": "Skipped duplicate rows"
          },
          "success_rate": {
            "type": "number",
            "title": "Success Rate",
            "description": "Success rate percentage"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ImportErrorDetail"
            },
            "type": "array",
            "title": "Errors",
            "description": "List of errors"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "List of warnings"
          }
        },
        "type": "object",
        "required": [
          "total_rows",
          "successful_rows",
          "failed_rows",
          "classified_rows",
          "skipped_duplicates",
          "success_rate"
        ],
        "title": "ImportResponse",
        "description": "Schema for import response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication (to be implemented)"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication (to be implemented)"
      }
    },
...
  },
...
}